:ukraine: Add initial setup for React application
This commit is the initial setup for the React application. It includes adding `package.json`, `package-lock.json`, `.gitignore`, `.eslintrc.cjs`, and `index.html`. It also adds the initial `App.tsx`,`main.tsx` and 'globals.css' files and their corresponding TypeScript and CSS files. It initializes eslint, TypeScript and dependencies required in package.json. Finally, it sets up the basic "Hello World" app in the root of the application.