:ukraine: Implement toast notifications and update user account management
This update includes the implementation of toast notifications for enhanced user experience across the application. Additionally, the functionality for user account creation has been advanced with the integration of Appwrite for better management of user data. Changes also include updates to the environment variables and dependencies in package.json to support these new features.