:ukraine: Introduce 'Get Recent Posts' and 'PostCard' features
The changes introduce a new feature for fetching recent posts and a 'PostCard' component for displaying individual posts. This involved updating 'queriesAndMutations' and 'api' modules for the fetching function and creating a new 'PostCard.tsx' file for the component. Various utility functions have been supplemented or altered to support these features as well.

Signed-off-by: diouf <geoffreyfleur@gmail.com>